requirements.txt
beautifulsoup4==4.12.3
easyocr==1.7.2
filelock==3.20.3
fsspec==2026.2.0
ImageIO==2.37.2
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
ninja==1.13.0
numpy==2.4.2
opencv-python-headless==4.13.0.92
packaging==26.0
pillow==12.1.1
pyclipper==1.4.0
python-bidi==0.6.7
PyYAML==6.0.3
python-dotenv==1.0.0
requests==2.32.3
scikit-image==0.26.0
scipy==1.17.0
setuptools==82.0.0
shapely==2.1.2
sympy==1.14.0
telethon==1.37.0
tifffile==2026.1.28
torch==2.10.0
torchvision==0.25.0
typing_extensions==4.15.0
